Add picnic-L3-FS [WIP]
This is an attempt to incorporate a Picnic instance in pqm4. Some
changes were carried out to make it compile:
  - Removed everything related to L1, L5 and -UR variants.
  - Removed some more functions in picnic.c that are not needed.
  - Assumed s == sm in stack.c. (That saves 76740 bytes of RAM -
  Otherwise this instance would never fit!)
  - Changed to use our randombytes.
  - Replaced SHA3 library with our fips202.
  - Removed calls to (f)printf/fflush.
  - Added api.h and crypto_sign_* wrappers.
  - Fixed some warnings.

I just noticed that I'm calling shake256_absorb wrongly. This is not meant
to be called sequentionally as it resets the Keccak state. Probably a
bunch of the absorb/squeezeblocks calls should then be replaced by a
regular shake256 call.
There are also still some mallocs left that should be removed.
For the test binary, only crypto_sign_keypair terminates.
Not sure whether stack overflow or whether I messed up some other
out-of-bounds thing.